+++ internal/project/creation.go
@@ -13,7 +13,7 @@ import (
 
 var ErrInvalidProjectTokenResponse = errors.New("invalid project access token response")
 
-func CreateNewProjectToken(gitlabClient *gitlab.Client, projectID int, entry *repository.Repository, prefix string) (*gitlab.ProjectAccessToken, error) {
+func CreateNewProjectToken(gitlabClient *gitlab.Client, projectID int64, entry *repository.Repository, prefix string) (*gitlab.ProjectAccessToken, error) {
 	l := logger.GetLogger()
 
 	l.Debug(fmt.Sprintf("Creating new project token for %s", *entry.RepoName))
@@ -39,7 +39,7 @@ func CreateNewProjectToken(gitlabClient *gitlab.Client, projectID int, entry *re
 	return token, nil
 }
 
-func createPATokenWithTokenOptions(gitlabClient *gitlab.Client, projectID int, name string, permissions []string, t *time.Time) (*gitlab.ProjectAccessToken, error) {
+func createPATokenWithTokenOptions(gitlabClient *gitlab.Client, projectID int64, name string, permissions []string, t *time.Time) (*gitlab.ProjectAccessToken, error) {
 	options := createProjectAccessTokenOptions(name, permissions, t)
 
 	token, _, err := gitlabClient.ProjectAccessTokens.CreateProjectAccessToken(projectID, options)
@@ -50,7 +50,7 @@ func createPATokenWithTokenOptions(gitlabClient *gitlab.Client, projectID int, n
 	return token, nil
 }
 
-func RenewProjectAccessToken(gitlabClient *gitlab.Client, projectID int, entry *repository.Repository, prefix string) (*gitlab.ProjectAccessToken, error) {
+func RenewProjectAccessToken(gitlabClient *gitlab.Client, projectID int64, entry *repository.Repository, prefix string) (*gitlab.ProjectAccessToken, error) {
 	tokenName, err := entry.NewTokenName(prefix)
 	if err != nil {
 		return nil, err
